[CK_BUILDER] ALMIOPEN-522: Testing-specific descriptor initialization
Remove old CK host descriptor helper dependency from CK Builder testing
framework and implement testing-owned descriptor computation.

Core changes (ALMIOPEN-522):
- Remove ck/library/utility/convolution_* includes from conv_fwd.hpp
- Add ConvFwdProblem struct (testing-owned conv parameter description)
- Add Args::make_conv_problem() (computes output spatial, no old CK dependency)
- Rewrite make_input/weight/output_descriptor() with testing-specific
  stride mapping for all supported layouts (GNHWC, NHWGC, etc.)
- Add optional explicit tensor-memory stride API (std::optional fields)
- Update conv_fwd_ck.hpp and conv_fwd_reference.hpp to use make_conv_problem()
- Remove to_ck_conv_param() method

Additional improvements:
- Add CK Tile forward EndToEnd support (conv_fwd_ck_tile.hpp + test)
- Proves shared Args/descriptor design works across backends
- Merge Create + EndToEnd tests into single file (matches old CK pattern)
- Fix unit_validation.cpp BF16 initialization for reliable testing

Results:
- Old CK warnings removed (no more RowMajor/ColumnMajor spam)
- All smoke tests pass (5/5)
- Old CK EndToEnd passes (XDL vs GPU reference)
- CK Tile EndToEnd passes (Tile vs GPU reference)

Note: This PR conflicts with #3518 (tile-builder-testing). Both touch
conv_fwd.hpp but with different approaches. This implementation directly
addresses ALMIOPEN-522 requirements by removing old CK dependency.

+++ experimental/builder/test/unit_validation.cpp
@@ -142,10 +142,12 @@ TEST(ValidationReportTests, MultipleSomeIncorrect)
         auto a = ckt::alloc_tensor_buffer(desc);
         auto b = ckt::alloc_tensor_buffer(desc);
 
-        ckt::fill_tensor_buffer(
-            desc, a.get(), [](size_t i) { return ck::type_convert<ck::bhalf_t>(i % 100); });
-        ckt::fill_tensor_buffer(
-            desc, b.get(), [](size_t i) { return ck::type_convert<ck::bhalf_t>(i % 101); });
+        ckt::fill_tensor_buffer(desc, a.get(), [](size_t i) {
+            return ck::type_convert<ck::bhalf_t>(static_cast<float>(i % 100));
+        });
+        ckt::fill_tensor_buffer(desc, b.get(), [](size_t i) {
+            return ck::type_convert<ck::bhalf_t>(static_cast<float>(i % 101));
+        });
 
         report.check("incorrect 1", desc, b.get(), a.get());
     }
